refactor: 사용 목적에 맞게 패키지 구조 변경
