--- model/src/main/kotlin/me/ogq/ocp/sample/model/Page.kt
@@ -1,6 +1,0 @@
-package me.ogq.ocp.sample.model
-
-data class Page(
-    val page: Int,
-    val pageSize: Int
-)
