--- model/src/main/kotlin/me/ogq/ocp/sample/model/SlicePage.kt
@@ -1,6 +1,0 @@
-package me.ogq.ocp.sample.model
-
-data class SlicePage<T>(
-    val hasNext: Boolean,
-    val elements: List<T>
-)
